backend/requirements.txt
fastapi==0.101.1
pydantic<2.0.0
requests
starlette
python-multipart
asyncpg
numpy
opencv-python-headless
alembic
SQLAlchemy
python-dotenv
redis
fastapi-cache2[redis]
PyJWT
bcrypt
pytz
Levenshtein
torch/torch-1.11.0+cpu-cp310-cp310-linux_x86_64.whl
torch/torchvision-0.12.0+cpu-cp310-cp310-linux_x86_64.whl
easyocr
